backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
pypdf==5.1.0
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-huggingface==0.1.2
langchain-groq==0.2.1
langchain-mongodb==0.3.0
pymongo==4.10.1
sentence-transformers==3.3.1
pydantic==2.10.1
httpx==0.27.2
